+++ tests/Resque/Tests/JobTest.php
@@ -425,8 +425,9 @@ class Some_Stub_Factory implements \Resque\Job\FactoryInterface
 
     /**
      * @param $className
-     * @param array $args
+     * @param $args
      * @param $queue
+     *
      * @return \Resque\Job\JobInterface
      */
     public function create($className, $args, $queue)